+++ src/core/hle/kernel/server_port.h
@@ -9,6 +9,7 @@
 #include <tuple>
 #include "common/common_types.h"
 #include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/wait_object.h"
 
 namespace Service {
 class SessionRequestHandler;
